+++ bigquery-connector-common/src/main/java/com/google/cloud/bigquery/connector/common/BigQueryClientFactory.java
@@ -1,18 +1,3 @@
-/*
- * Copyright 2018 Google Inc. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
 package com.google.cloud.bigquery.connector.common;
 
 import com.google.api.gax.core.FixedCredentialsProvider;
@@ -20,21 +5,24 @@ import com.google.api.gax.grpc.InstantiatingGrpcChannelProvider;
 import com.google.auth.Credentials;
 import com.google.cloud.bigquery.storage.v1.BigQueryReadClient;
 import com.google.cloud.bigquery.storage.v1.BigQueryReadSettings;
+import com.google.cloud.bigquery.storage.v1beta2.BigQueryWriteClient;
+import com.google.cloud.bigquery.storage.v1beta2.BigQueryWriteSettings;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import javax.inject.Inject;
 import java.io.IOException;
 import java.io.Serializable;
 import java.io.UncheckedIOException;
 import java.util.Optional;
-import javax.inject.Inject;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 /**
  * Since Guice recommends to avoid injecting closeable resources (see
  * https://github.com/google/guice/wiki/Avoid-Injecting-Closable-Resources), this factory creates
  * short lived clients that can be closed independently.
  */
-public class BigQueryReadClientFactory implements Serializable {
-  private static final Logger log = LoggerFactory.getLogger(BigQueryReadClientFactory.class);
+public class BigQueryClientFactory implements Serializable {
+  private static final Logger log = LoggerFactory.getLogger(BigQueryClientFactory.class);
 
   private final Credentials credentials;
   // using the user agent as HeaderProvider is not serializable
@@ -42,7 +30,7 @@ public class BigQueryReadClientFactory implements Serializable {
   private final BigQueryConfig bqConfig;
 
   @Inject
-  public BigQueryReadClientFactory(
+  public BigQueryClientFactory(
       BigQueryCredentialsSupplier bigQueryCredentialsSupplier,
       UserAgentHeaderProvider userAgentHeaderProvider,
       BigQueryConfig bqConfig) {
@@ -57,14 +45,7 @@ public class BigQueryReadClientFactory implements Serializable {
       InstantiatingGrpcChannelProvider.Builder transportBuilder =
           BigQueryReadSettings.defaultGrpcTransportProviderBuilder()
               .setHeaderProvider(userAgentHeaderProvider);
-      BigQueryProxyConfig proxyConfig = bqConfig.getBigQueryProxyConfig();
-      if (proxyConfig.getProxyUri().isPresent()) {
-        transportBuilder.setChannelConfigurator(
-            BigQueryProxyTransporterBuilder.createGrpcChannelConfigurator(
-                proxyConfig.getProxyUri(),
-                proxyConfig.getProxyUsername(),
-                proxyConfig.getProxyPassword()));
-      }
+      setProxyConfig(transportBuilder);
       endpoint.ifPresent(
           e -> {
             log.info("Overriding endpoint to: ", e);
@@ -76,7 +57,34 @@ public class BigQueryReadClientFactory implements Serializable {
               .setCredentialsProvider(FixedCredentialsProvider.create(credentials));
       return BigQueryReadClient.create(clientSettings.build());
     } catch (IOException e) {
-      throw new UncheckedIOException("Error creating BigQueryStorageClient", e);
+      throw new UncheckedIOException("Error creating BigQueryStorageReadClient", e);
+    }
+  }
+
+  public BigQueryWriteClient createBigQueryWriteClient() {
+    try {
+      InstantiatingGrpcChannelProvider.Builder transportBuilder =
+          BigQueryWriteSettings.defaultGrpcTransportProviderBuilder()
+              .setHeaderProvider(userAgentHeaderProvider);
+      setProxyConfig(transportBuilder);
+      BigQueryWriteSettings.Builder clientSettings =
+          BigQueryWriteSettings.newBuilder()
+              .setTransportChannelProvider(transportBuilder.build())
+              .setCredentialsProvider(FixedCredentialsProvider.create(credentials));
+      return BigQueryWriteClient.create(clientSettings.build());
+    } catch (IOException e) {
+      throw new BigQueryConnectorException("Error creating BigQueryWriteClient", e);
+    }
+  }
+
+  public void setProxyConfig(InstantiatingGrpcChannelProvider.Builder transportBuilder) {
+    BigQueryProxyConfig proxyConfig = bqConfig.getBigQueryProxyConfig();
+    if (proxyConfig.getProxyUri().isPresent()) {
+      transportBuilder.setChannelConfigurator(
+          BigQueryProxyTransporterBuilder.createGrpcChannelConfigurator(
+              proxyConfig.getProxyUri(),
+              proxyConfig.getProxyUsername(),
+              proxyConfig.getProxyPassword()));
     }
   }
 }
