+++ bigquery-connector-common/src/main/java/com/google/cloud/bigquery/connector/common/BigQueryClientModule.java
@@ -34,11 +34,10 @@ public class BigQueryClientModule implements com.google.inject.Module {
   @Override
   public void configure(Binder binder) {
     // BigQuery related
-    binder.bind(BigQueryReadClientFactory.class).in(Scopes.SINGLETON);
+    binder.bind(BigQueryClientFactory.class).in(Scopes.SINGLETON);
     binder
         .bind(BigQueryTracerFactory.class)
         .toInstance(new LoggingBigQueryTracerFactory(/*Log every 2^14 batches*/ 14));
-    binder.bind(BigQueryWriteClientFactory.class).in(Scopes.SINGLETON);
   }
 
   @Provides
