+++ bigquery-connector-common/src/main/java/com/google/cloud/bigquery/connector/common/BigQueryErrorCode.java
@@ -19,6 +19,7 @@ public enum BigQueryErrorCode {
   BIGQUERY_VIEW_DESTINATION_TABLE_CREATION_FAILED(0),
   BIGQUERY_DATETIME_PARSING_ERROR(1),
   BIGQUERY_FAILED_TO_EXECUTE_QUERY(2),
+  BIGQUERY_INVALID_SCHEMA(2),
   // Should be last
   UNSUPPORTED(9998),
   UNKNOWN(9999);
