+++ spark-bigquery-dsv2/spark-bigquery-spark24/src/main/java/com/google/cloud/spark/bigquery/v2/ArrowInputPartition.java
@@ -16,8 +16,7 @@
 package com.google.cloud.spark.bigquery.v2;
 
 import static com.google.common.base.Optional.fromJavaUtil;
-
-import com.google.cloud.bigquery.connector.common.BigQueryReadClientFactory;
+import com.google.cloud.bigquery.connector.common.BigQueryClientFactory;
 import com.google.cloud.bigquery.connector.common.BigQueryStorageReadRowsTracer;
 import com.google.cloud.bigquery.connector.common.BigQueryTracerFactory;
 import com.google.cloud.bigquery.connector.common.ReadRowsHelper;
@@ -38,7 +37,7 @@ import org.apache.spark.sql.vectorized.ColumnarBatch;
 
 public class ArrowInputPartition implements InputPartition<ColumnarBatch> {
 
-  private final BigQueryReadClientFactory bigQueryReadClientFactory;
+  private final BigQueryClientFactory bigQueryReadClientFactory;
   private final BigQueryTracerFactory tracerFactory;
   private final List<String> streamNames;
   private final ReadRowsHelper.Options options;
@@ -47,7 +46,7 @@ public class ArrowInputPartition implements InputPartition<ColumnarBatch> {
   private final com.google.common.base.Optional<StructType> userProvidedSchema;
 
   public ArrowInputPartition(
-      BigQueryReadClientFactory bigQueryReadClientFactory,
+      BigQueryClientFactory bigQueryReadClientFactory,
       BigQueryTracerFactory tracerFactory,
       List<String> names,
       ReadRowsHelper.Options options,
