+++ spark-bigquery-dsv2/spark-bigquery-spark24/src/main/java/com/google/cloud/spark/bigquery/v2/BigQueryDataSourceReaderModule.java
@@ -17,7 +17,7 @@ package com.google.cloud.spark.bigquery.v2;
 
 import com.google.cloud.bigquery.TableInfo;
 import com.google.cloud.bigquery.connector.common.BigQueryClient;
-import com.google.cloud.bigquery.connector.common.BigQueryReadClientFactory;
+import com.google.cloud.bigquery.connector.common.BigQueryClientFactory;
 import com.google.cloud.bigquery.connector.common.BigQueryTracerFactory;
 import com.google.cloud.spark.bigquery.SparkBigQueryConfig;
 import com.google.inject.Binder;
@@ -36,7 +36,7 @@ public class BigQueryDataSourceReaderModule implements Module {
   @Provides
   public BigQueryDataSourceReader provideDataSourceReader(
       BigQueryClient bigQueryClient,
-      BigQueryReadClientFactory bigQueryReadClientFactory,
+      BigQueryClientFactory bigQueryReadClientFactory,
       BigQueryTracerFactory tracerFactory,
       SparkBigQueryConfig config,
       SparkSession sparkSession) {
