+++ spark-bigquery-dsv2/spark-bigquery-spark24/src/main/java/com/google/cloud/spark/bigquery/v2/BigQueryDataSourceV2.java
@@ -45,6 +45,27 @@ import org.apache.spark.sql.types.StructType;
 public class BigQueryDataSourceV2
     implements DataSourceV2, DataSourceRegister, ReadSupport, WriteSupport {
 
+  private enum WriteMethod {
+    DIRECT("direct"),
+    INDIRECT("indirect");
+
+    private final String writePath;
+
+    WriteMethod(String writePath) {
+      this.writePath = writePath;
+    }
+
+    static WriteMethod getWriteMethod(Optional<String> path) {
+      if (!path.isPresent() || path.get().equalsIgnoreCase("direct")) {
+        return DIRECT;
+      } else if (path.get().equalsIgnoreCase("indirect")) {
+        return INDIRECT;
+      } else {
+        throw new IllegalArgumentException("Unknown writePath Provided for writing the DataFrame");
+      }
+    }
+  }
+
   @Override
   public DataSourceReader createReader(StructType schema, DataSourceOptions options) {
     Injector injector = createInjector(schema, options, new BigQueryDataSourceReaderModule());
@@ -81,12 +102,12 @@ public class BigQueryDataSourceV2
   @Override
   public Optional<DataSourceWriter> createWriter(
       String writeUUID, StructType schema, SaveMode mode, DataSourceOptions options) {
-    String path = options.get("writePath").orElse("direct");
+    WriteMethod path = WriteMethod.getWriteMethod(options.get("writePath"));
 
-    if (path.equalsIgnoreCase("direct")) {
+    if (path.equals(WriteMethod.DIRECT)) {
       return createDirectDataSourceWriter(writeUUID, schema, mode, options);
-    } else if (path.equalsIgnoreCase("inDirect")) {
-      return createInDirectDataSourceWriter(writeUUID, schema, mode, options);
+    } else if (path.equals(WriteMethod.INDIRECT)) {
+      return createIndirectDataSourceWriter(writeUUID, schema, mode, options);
     } else {
       throw new IllegalArgumentException("Unknown writePath Provided for writing the DataFrame");
     }
@@ -103,7 +124,7 @@ public class BigQueryDataSourceV2
     return Optional.of(writer);
   }
 
-  private Optional<DataSourceWriter> createInDirectDataSourceWriter(
+  private Optional<DataSourceWriter> createIndirectDataSourceWriter(
       String writeUUID, StructType schema, SaveMode mode, DataSourceOptions options) {
     Injector injector =
         createInjector(
