+++ spark-bigquery-dsv2/spark-bigquery-spark24/src/test/java/com/google/cloud/spark/bigquery/integration/Spark24WriteIntegrationTest.java
@@ -17,30 +17,34 @@ package com.google.cloud.spark.bigquery.integration;
 
 import com.google.cloud.RetryOption;
 import com.google.cloud.ServiceOptions;
-import com.google.cloud.bigquery.*;
+import com.google.cloud.bigquery.BigQuery;
+import com.google.cloud.bigquery.BigQueryOptions;
+import com.google.cloud.bigquery.DatasetInfo;
+import com.google.cloud.bigquery.Job;
+import com.google.cloud.bigquery.JobInfo;
+import com.google.cloud.bigquery.QueryJobConfiguration;
+import com.google.cloud.bigquery.TableId;
+import com.google.cloud.bigquery.TableInfo;
 import org.apache.log4j.Level;
 import org.apache.log4j.LogManager;
 import org.apache.log4j.Logger;
-import org.apache.spark.bigquery.BigNumeric;
-import org.apache.spark.bigquery.BigQueryDataTypes;
-import org.apache.spark.sql.*;
 import org.apache.spark.sql.Dataset;
-import org.apache.spark.sql.types.*;
+import org.apache.spark.sql.Row;
+import org.apache.spark.sql.SaveMode;
+import org.apache.spark.sql.SparkSession;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import org.threeten.bp.Duration;
 
 import java.io.IOException;
-import java.math.BigDecimal;
-import java.math.MathContext;
-import java.sql.Date;
-import java.sql.Timestamp;
 import java.util.Arrays;
 
+import static com.google.cloud.spark.bigquery.integration.TestConstants.STORAGE_API_ALL_TYPES_ROWS;
+import static com.google.cloud.spark.bigquery.integration.TestConstants.STORAGE_API_ALL_TYPES_SCHEMA;
+import static com.google.cloud.spark.bigquery.integration.TestConstants.STORAGE_API_ALL_TYPES_SCHEMA_BIGQUERY_REPRESENTATION;
 import static com.google.common.truth.Truth.assertThat;
 import static java.lang.String.format;
-import static org.apache.spark.sql.types.DataTypes.*;
 import static org.junit.Assert.fail;
 
 public class Spark24WriteIntegrationTest extends WriteIntegrationTestBase {
@@ -51,11 +55,6 @@ public class Spark24WriteIntegrationTest extends WriteIntegrationTestBase {
 
   // Numeric is a fixed precision Decimal Type with 38 digits of precision and 9 digits of scale.
   // See https://cloud.google.com/bigquery/docs/reference/standard-sql/data-types#numeric-type
-  private static final int BQ_NUMERIC_PRECISION = 38;
-  private static final int BQ_NUMERIC_SCALE = 9;
-  private static final DecimalType NUMERIC_SPARK_TYPE =
-      DataTypes.createDecimalType(BQ_NUMERIC_PRECISION, BQ_NUMERIC_SCALE);
-
   public static final Logger logger = LogManager.getLogger("com.google.cloud");
 
   public static final String PROJECT = ServiceOptions.getDefaultProjectId();
@@ -75,18 +74,6 @@ public class Spark24WriteIntegrationTest extends WriteIntegrationTestBase {
   public static final String MB100_TABLE = "annotations_bbox";
   public static final String MB100_ID =
       BIGQUERY_PUBLIC_DATA + ":" + MB100_DATASET + "." + MB100_TABLE; // 156 MB
-  public static final String GB3_DATASET = "open_images";
-  public static final String GB3_TABLE = "images";
-  public static final String GB3_ID =
-      BIGQUERY_PUBLIC_DATA + ":" + GB3_DATASET + "." + GB3_TABLE; // 3.56 GB
-  public static final String GB20_DATASET = "samples";
-  public static final String GB20_TABLE = "natality";
-  public static final String GB20_ID =
-      BIGQUERY_PUBLIC_DATA + ":" + GB20_DATASET + "." + GB20_TABLE; // 21 GB
-  public static final String GB250_DATASET = "samples";
-  public static final String GB250_TABLE = "trigrams";
-  public static final String GB250_ID =
-      BIGQUERY_PUBLIC_DATA + ":" + GB250_DATASET + "." + GB250_TABLE; // 256 GB
 
   public static SparkSession spark;
   public static BigQuery bigquery;
@@ -96,9 +83,6 @@ public class Spark24WriteIntegrationTest extends WriteIntegrationTestBase {
   public static Dataset<Row> smallDataDf;
   public static Dataset<Row> MB20Df;
   public static Dataset<Row> MB100Df;
-  public static Dataset<Row> GB3Df;
-  public static Dataset<Row> GB20Df;
-  public static Dataset<Row> GB250Df;
 
   @BeforeClass
   public static void init() throws Exception {
@@ -109,7 +93,9 @@ public class Spark24WriteIntegrationTest extends WriteIntegrationTestBase {
             .config("some-config", "some-value")
             .master("local[*]")
             .getOrCreate();
-    allTypesDf = spark.createDataFrame(Arrays.asList(ALL_TYPES_ROWS), ALL_TYPES_SCHEMA);
+    allTypesDf =
+        spark.createDataFrame(
+            Arrays.asList(STORAGE_API_ALL_TYPES_ROWS), STORAGE_API_ALL_TYPES_SCHEMA);
     smallDataDf = spark.read().format("bigquery").option("table", SMALL_DATA_ID).load();
     twiceAsBigDf = smallDataDf.unionAll(smallDataDf);
     MB20Df =
@@ -124,11 +110,6 @@ public class Spark24WriteIntegrationTest extends WriteIntegrationTestBase {
             .toDF();
     MB100Df =
         spark.read().format("bigquery").option("table", MB100_ID).load() /*.coalesce(20).toDF()*/;
-    GB3Df = spark.read().format("bigquery").option("table", GB3_ID).load() /*.coalesce(20).toDF()*/;
-    GB20Df =
-        spark.read().format("bigquery").option("table", GB20_ID).load() /*.coalesce(20).toDF()*/;
-    GB250Df =
-        spark.read().format("bigquery").option("table", GB250_ID).load() /*.coalesce(20).toDF()*/;
 
     bigquery = BigQueryOptions.getDefaultInstance().getService();
     DatasetInfo datasetInfo =
@@ -213,7 +194,7 @@ public class Spark24WriteIntegrationTest extends WriteIntegrationTestBase {
             .option("project", PROJECT)
             .load();
 
-    assertThat(actualDF.schema()).isEqualTo(ALL_TYPES_SCHEMA_BIGQUERY_REPRESENTATION);
+    assertThat(actualDF.schema()).isEqualTo(STORAGE_API_ALL_TYPES_SCHEMA_BIGQUERY_REPRESENTATION);
 
     Dataset<Row> intersection = actualDF.intersectAll(expectedDF);
     assertThat(
@@ -398,166 +379,4 @@ public class Spark24WriteIntegrationTest extends WriteIntegrationTestBase {
                     .getTable(TableId.of(BIGQUERY_PUBLIC_DATA, MB100_DATASET, MB100_TABLE))
                     .getNumBytes()));
   }
-
-  public static final StructType ALL_TYPES_SCHEMA =
-      new StructType()
-          .add(
-              new StructField(
-                  "int_req",
-                  IntegerType,
-                  false,
-                  new MetadataBuilder().putString("description", "required integer").build()))
-          .add(new StructField("int_null", IntegerType, true, Metadata.empty()))
-          .add(new StructField("long", LongType, true, Metadata.empty()))
-          .add(new StructField("short", ShortType, true, Metadata.empty()))
-          .add(new StructField("bytenum", ByteType, true, Metadata.empty()))
-          .add(new StructField("bool", BooleanType, true, Metadata.empty()))
-          .add(new StructField("str", StringType, true, Metadata.empty()))
-          .add(new StructField("date", DateType, true, Metadata.empty()))
-          .add(new StructField("timestamp", TimestampType, true, Metadata.empty()))
-          .add(new StructField("binary", BinaryType, true, Metadata.empty()))
-          .add(new StructField("float", DoubleType, true, Metadata.empty()))
-          .add(
-              new StructField(
-                  "nums",
-                  new StructType()
-                      .add(new StructField("min", NUMERIC_SPARK_TYPE, true, Metadata.empty()))
-                      .add(new StructField("max", NUMERIC_SPARK_TYPE, true, Metadata.empty()))
-                      .add(new StructField("pi", NUMERIC_SPARK_TYPE, true, Metadata.empty()))
-                      .add(new StructField("big_pi", NUMERIC_SPARK_TYPE, true, Metadata.empty())),
-                  true,
-                  Metadata.empty()))
-          .add(
-              new StructField(
-                  "big_numeric_nums",
-                  new StructType()
-                      .add(
-                          new StructField(
-                              "min", BigQueryDataTypes.BigNumericType, true, Metadata.empty()))
-                      .add(
-                          new StructField(
-                              "max", BigQueryDataTypes.BigNumericType, true, Metadata.empty()))
-                      .add(
-                          new StructField(
-                              "pi", BigQueryDataTypes.BigNumericType, true, Metadata.empty())),
-                  true,
-                  Metadata.empty()))
-          .add(new StructField("int_arr", new ArrayType(IntegerType, true), true, Metadata.empty()))
-          .add(
-              new StructField(
-                  "int_struct_arr",
-                  new ArrayType(
-                      new StructType()
-                          .add(new StructField("i", IntegerType, true, Metadata.empty())),
-                      true),
-                  true,
-                  Metadata.empty()));
-
-  // same as ALL_TYPES_SCHEMA, except all IntegerType's are LongType's.
-  public static final StructType ALL_TYPES_SCHEMA_BIGQUERY_REPRESENTATION =
-      new StructType()
-          .add(
-              new StructField(
-                  "int_req",
-                  LongType,
-                  false,
-                  new MetadataBuilder()
-                      .putString("description", "required integer")
-                      .putString("comment", "required integer")
-                      .build()))
-          .add(new StructField("int_null", LongType, true, Metadata.empty()))
-          .add(new StructField("long", LongType, true, Metadata.empty()))
-          .add(new StructField("short", LongType, true, Metadata.empty()))
-          .add(new StructField("bytenum", LongType, true, Metadata.empty()))
-          .add(new StructField("bool", BooleanType, true, Metadata.empty()))
-          .add(new StructField("str", StringType, true, Metadata.empty()))
-          .add(new StructField("date", DateType, true, Metadata.empty()))
-          .add(new StructField("timestamp", TimestampType, true, Metadata.empty()))
-          .add(new StructField("binary", BinaryType, true, Metadata.empty()))
-          .add(new StructField("float", DoubleType, true, Metadata.empty()))
-          .add(
-              new StructField(
-                  "nums",
-                  new StructType()
-                      .add(new StructField("min", NUMERIC_SPARK_TYPE, true, Metadata.empty()))
-                      .add(new StructField("max", NUMERIC_SPARK_TYPE, true, Metadata.empty()))
-                      .add(new StructField("pi", NUMERIC_SPARK_TYPE, true, Metadata.empty()))
-                      .add(new StructField("big_pi", NUMERIC_SPARK_TYPE, true, Metadata.empty())),
-                  true,
-                  Metadata.empty()))
-          .add(
-              new StructField(
-                  "big_numeric_nums",
-                  new StructType()
-                      .add(
-                          new StructField(
-                              "min", BigQueryDataTypes.BigNumericType, true, Metadata.empty()))
-                      .add(
-                          new StructField(
-                              "max", BigQueryDataTypes.BigNumericType, true, Metadata.empty()))
-                      .add(
-                          new StructField(
-                              "pi", BigQueryDataTypes.BigNumericType, true, Metadata.empty())),
-                  true,
-                  Metadata.empty()))
-          .add(new StructField("int_arr", new ArrayType(LongType, true), true, Metadata.empty()))
-          .add(
-              new StructField(
-                  "int_struct_arr",
-                  new ArrayType(
-                      new StructType().add(new StructField("i", LongType, true, Metadata.empty())),
-                      true),
-                  true,
-                  Metadata.empty()));
-
-  public static final Row[] ALL_TYPES_ROWS =
-      new Row[] {
-        RowFactory.create(
-            123456789,
-            null,
-            123456789L,
-            (short) 1024,
-            (byte) 127,
-            true,
-            "hello",
-            Date.valueOf("2019-03-18"),
-            new Timestamp(1552872225000L), // 2019-03-18 01:23:45
-            new byte[] {
-              98, 121, 116, 101, 115
-            }, // byte[] representation of string "bytes" -> stored in BQ as Ynl0ZXM=
-            1.2345,
-            RowFactory.create(
-                Decimal.apply(
-                    new BigDecimal(
-                        "-99999999999999999999999999999.999999999",
-                        new MathContext(BQ_NUMERIC_PRECISION)),
-                    BQ_NUMERIC_PRECISION,
-                    BQ_NUMERIC_SCALE),
-                Decimal.apply(
-                    new BigDecimal(
-                        "99999999999999999999999999999.999999999",
-                        new MathContext(BQ_NUMERIC_PRECISION)),
-                    BQ_NUMERIC_PRECISION,
-                    BQ_NUMERIC_SCALE),
-                Decimal.apply(
-                    new BigDecimal("3.14", new MathContext(BQ_NUMERIC_PRECISION)),
-                    BQ_NUMERIC_PRECISION,
-                    BQ_NUMERIC_SCALE),
-                Decimal.apply(
-                    new BigDecimal(
-                        "31415926535897932384626433832.795028841",
-                        new MathContext(BQ_NUMERIC_PRECISION)),
-                    BQ_NUMERIC_PRECISION,
-                    BQ_NUMERIC_SCALE)),
-            RowFactory.create(
-                new BigNumeric(
-                    new BigDecimal(
-                        "-578960446186580977117854925043439539266.34992332820282019728792003956564819968")),
-                new BigNumeric(
-                    new BigDecimal(
-                        "578960446186580977117854925043439539266.34992332820282019728792003956564819967")),
-                new BigNumeric(new BigDecimal("3.14"))),
-            new int[] {1, 2, 3, 4},
-            new Row[] {RowFactory.create(1), RowFactory.create(1)})
-      };
 }
